Adding changes

+++ cloud/amazon/resources/keypair.go
@@ -21,6 +21,7 @@ import (
 	"github.com/kris-nova/kubicorn/cloud"
 	"github.com/kris-nova/kubicorn/cutil/compare"
 	"github.com/kris-nova/kubicorn/cutil/logger"
+	"strings"
 )
 
 type KeyPair struct {
@@ -67,7 +68,10 @@ func (r *KeyPair) Actual(known *cluster.Cluster) (cloud.Resource, error) {
 		actual.CloudID = *keypair.KeyName
 		actual.PublicKeyFingerprint = *keypair.KeyFingerprint
 	}
-	actual.User = known.SSH.User
+	actual.PublicKeyPath = known.Ssh.PublicKeyPath
+	actual.PublicKeyData = string(known.Ssh.PublicKeyData)
+	actual.PublicKeyFingerprint = known.Ssh.PublicKeyFingerprint
+	actual.User = known.Ssh.User
 	r.CachedActual = actual
 	return actual, nil
 }
@@ -109,36 +113,41 @@ func (r *KeyPair) Apply(actual, expected cloud.Resource, applyCluster *cluster.C
 		KeyName:           &expected.(*KeyPair).Name,
 		PublicKeyMaterial: []byte(expected.(*KeyPair).PublicKeyData),
 	}
+	newResource := &KeyPair{}
 	output, err := Sdk.Ec2.ImportKeyPair(input)
 	if err != nil {
-		return nil, err
+		if !strings.Contains(err.Error(), "InvalidKeyPair.Duplicate") {
+			return nil, err
+		}
+		logger.Info("Using existing KeyPair [%s]", expected.(*KeyPair).Name)
+	} else {
+		logger.Info("Created KeyPair [%s]", *output.KeyName)
+		newResource.PublicKeyFingerprint = *output.KeyFingerprint
 	}
-	logger.Info("Created KeyPair [%s]", *output.KeyName)
-
-	newResource := &KeyPair{}
+	newResource.CloudID = expected.(*KeyPair).Name
 	newResource.PublicKeyData = expected.(*KeyPair).PublicKeyData
 	newResource.PublicKeyPath = expected.(*KeyPair).PublicKeyPath
 	newResource.User = expected.(*KeyPair).User
-	newResource.PublicKeyFingerprint = *output.KeyFingerprint
-	newResource.CloudID = expected.(*KeyPair).Name
 	newResource.Name = expected.(*KeyPair).Name
 	return newResource, nil
 }
 func (r *KeyPair) Delete(actual cloud.Resource, known *cluster.Cluster) (cloud.Resource, error) {
 	logger.Debug("keypair.Delete")
-	deleteResource := actual.(*KeyPair)
-	if deleteResource.CloudID == "" {
-		return nil, fmt.Errorf("Unable to delete keypair resource without ID [%s]", deleteResource.Name)
+	force := false
+	if force {
+		deleteResource := actual.(*KeyPair)
+		if deleteResource.CloudID == "" {
+			return nil, fmt.Errorf("Unable to delete keypair resource without ID [%s]", deleteResource.Name)
+		}
+		input := &ec2.DeleteKeyPairInput{
+			KeyName: &actual.(*KeyPair).Name,
+		}
+		_, err := Sdk.Ec2.DeleteKeyPair(input)
+		if err != nil {
+			return nil, err
+		}
+		logger.Info("Deleted keypair [%s]", actual.(*KeyPair).CloudID)
 	}
-
-	input := &ec2.DeleteKeyPairInput{
-		KeyName: &actual.(*KeyPair).Name,
-	}
-	_, err := Sdk.Ec2.DeleteKeyPair(input)
-	if err != nil {
-		return nil, err
-	}
-	logger.Info("Deleted keypair [%s]", actual.(*KeyPair).CloudID)
 	newResource := &KeyPair{}
 	newResource.Tags = actual.(*KeyPair).Tags
 	newResource.Name = actual.(*KeyPair).Name
